+++ ReleaseNotes.md
@@ -1,3 +1,7 @@
+### New in 0.0.5 (Released 2015/10/10)
+* Add AvailabilityZone aliases
+* Get MetaData from EC2 nuget
+
 ### New in 0.0.4 (Released 2015/09/28)
 * Add Namespace Categories
 * Remove Cake.Core reference